prepared models for uTorrent. added ability to open the webGUI with a long tap on the cell

+++ TorrentAdder/SYAdder.m
@@ -127,7 +127,7 @@
 {
     [self setResponse:response forConnection:connection];
     
-    int code = [(NSHTTPURLResponse*)response statusCode];
+    NSInteger code = [(NSHTTPURLResponse*)response statusCode];
     if(code == 409) {
         NSMutableURLRequest *newRequest = [[connection originalRequest] mutableCopy];
         NSString *sessionID = [[(NSHTTPURLResponse*)response allHeaderFields]
